Latest debug code showing problem
Latest debug which shows that the unit test testStartVolumeGroup()
hanging.   The hang is due to the watcher NOT seeing the VG transition
to ONLINING.   vgMap lists the VG as ONLINING when we have not yet
received the watch event.  The watch event is ignored because it is for
the VG to go to ONLINING.

The problem is vgMap is populated after the txn() to move it to ONLINING
but before the watch event is received.

vgMap is there to handle cases where we cannot retrieve the revision of
the log due to compaction.   There is a bit of a catch 22 here.

Anyway, save this reproduction case for now while working on etcd-mgmt.