Introduced new FetchExtentMapChunk API (JSON RPC callable)
This new API replaces, in most cases, GetReadPlan that was designed to be
used when a known range of a FileInode is to be read and there is no need
to limit the number of resultant extents that are returned.

New users should instead adopt FetchExtentMapChunk and implement a "paging"
model where only a bounded number of extents are returned in each request.