Introduced fileInode references field
This will be used to safely track InodeCache contents and is the building
block for managing FileInode leases and locks.

Note that inode.FetchExtentMapChunk() and friends now also
returns the FileInode's FileSize. This is helpful particularly
on the first request for a FileInode so that pfsagentd can
accurately represent all three logical extent types:

  Data - found in LogSegments
  Zero-filled holes
  Unfetched ExtentMap ranges